Migrate monetary fields from double to BigDecimal
Currency amounts stored as double risk silent rounding drift, which
matters once ledger reconciliation depends on exact balances.

- Entities (GiftCard, GiftCardHold, LedgerEntry) and DTOs now use
  BigDecimal for balance/amount fields.
- Services use compareTo/add/subtract instead of primitive operators.
- New Flyway migration (V15) alters existing DOUBLE PRECISION columns
  to NUMERIC(19,2).
- Idempotency-key hashing normalizes amounts to a fixed 2-decimal
  scale before hashing, so equivalent amounts with different JSON
  representations still hit the same cache key.
- Updated unit and integration tests accordingly.

+++ src/main/java/com/finovago/p2p/service/LedgerService.java
@@ -1,5 +1,7 @@
 package com.finovago.p2p.service;
 
+import java.math.BigDecimal;
+
 import org.springframework.stereotype.Service;
 
 import com.finovago.p2p.model.GiftCard;
@@ -20,7 +22,7 @@ public class LedgerService {
      * Appends a ledger row. Never call this outside the same transaction as the balance
      * mutation it records - the two must commit or roll back together.
      */
-    public void record(GiftCard giftCard, Long merchantId, LedgerEntryType entryType, double amount, double balanceAfter, Long referenceId) {
+    public void record(GiftCard giftCard, Long merchantId, LedgerEntryType entryType, BigDecimal amount, BigDecimal balanceAfter, Long referenceId) {
         ledgerEntryRepository.save(new LedgerEntry(giftCard, merchantId, entryType, amount, balanceAfter, referenceId));
     }
 }
